egl-wayland: Exit early from eglGetPlatformDisplay for non-NV devices
In wlEglGetPlatformDisplayExport, after we get the DRM device node from
the server, call drmGetDevice to check if it points to an NVIDIA device.

If it's not an NVIDIA device, then return early before calling
eglQueryDevicesEXT.

If the display is configured to use GPU offloading (either with
EGL_EXT_explicit_device or if __NV_PRIME_RENDER_OFFLOAD is set), then it
will still continue and look for a matching/NVIDIA device regardless of
what the server is running.

On multi-GPU systems, calling eglQueryDevicesEXT can cause it to power
the GPU's on, which can take a very long time.