egl-wayland: Use the known versions of Wayland protocols
Currently, we bind the latest versions of the Wayland protocols the
compositors implement. However, this is not always correct. For example,
a new version of a protocol we use could add new events, which we would
be requesting by binding the newer version of the protocol, but we
wouldn't be handling because the event simply didn't exist when we
started using the protocol. This could lead to protocol errors. So, it
is important that we bind the latest versions of the protocols that we
implemented the required callbacks for and tested instead of the latest
versions the compositors implement.