egl-wayland: Use a swapchain of 4 images for Wayland
This fixes KDE freezing when explicit sync is in use.

The problem users are running into is that we always have to advance to
a new buffer because the driver always expects to be incremented to the
next valid buffer as part of swapbuffers. So currently it seems one of
the three images will always be owned by the driver (either the buffer
currently/just rendered to, or the one we just advanced to for future
rendering)

So the three buffers are used up by:
  1. One buffer owned by the driver
  2. One buffer that just got committed and shared with the compositor
  3. One buffer owned by the compositor, pending a release

For whatever reason Kwin is holding onto 2 and 3 indefinitely when the
dock gets hidden, and we hold onto 1 and try waiting for one of the
other two to become free. We need a fourth to allow us to continue
feeding the driver.

The easiest fix for this is to use four buffers instead of three. This
is done to better match what Mesa does, as apparently 4 is the
expectation on wayland.
https://gitlab.freedesktop.org/mesa/mesa/-/issues/6249#note_1328923

We only wait indefinitely when all but one buffers are pending. There are
four total buffers:
  1. One owned by the driver (as per above)
  2. One we just committed and sent to the compositor
  3. One owned by compositor, queued for scanout
  4. One owned by compositor, in process of releasing

Not all compositors will hold 3 and 4 indefinitely, although Kwin does
at certain times.