egl-wayland: Use the correct coordinates when setting the damage region
Using the buffer width and height with wl_surface_damage is incorrect, as it expects surface-local coordinates. This causes issues when scaling using a viewport with an output region larger than the source region, as only a portion of the output region equal to the unscaled buffer size will be updated.

Use wl_surface_damage_buffer, when available, to set the damage region using buffer-local coordinates, otherwise, use wl_surface_damage with a sufficiently large width and height to declare the entire output region as damaged.