egl-wayland: check for driver explicit sync support
This uses some error conditions added by CL 33837064 to check if
we can actually do explicit sync or if we should fall back. The problem
is that we currently do not have a good version detection mechanism
available for us to do this with, so we give EGL an invalid fd+status
combo to trigger a condition that was added by the same CL as what
we need.