egl-wayland: Initialize streamImages during Pbuffer surface creation
We are missing the initializer for the streamImages wl_list in
wlEglCreatePbufferSurfaceHook. This ensures the list is valid when we
eventually destroy the surface and have to check the list contents.

This change adds a common initialization function
wlEglInitializeSurfaceCommon which is used by both surface creators
to set up common state such as these wl_lists.

This fixes crashes with qt6webengine.