Add a hook for eglQueryString.
Add a hook function for eglQueryString, so that we can append
EGL_EXT_present_opaque to the extension string to a Wayland EGLDisplay.

Remove EGL_EXT_present_opaque from wlEglQueryStringExport, because we
don't want to add that extension to non-Wayland EGLDisplays.

Note that this requires a more recent driver that correctly calls the
eglQueryString hook.